crypto-portfolio-tracker-backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
python-multipart